Fixed issues with opening chests
When a weapon is active, there end up being a lot of wackiness with opening chest inventories, added code to allow server inventory packets to get through when a player opens a chest even if the weapon item is currently active.
