+++ run/world/stats/0e35bb95-948e-4baf-813c-5e285bfda781.json
@@ -1,0 +1,1 @@
+{"stats":{"minecraft:picked_up":{"minecraft:rotten_flesh":19},"minecraft:custom":{"minecraft:jump":355,"minecraft:time_since_rest":1629,"minecraft:crouch_one_cm":12,"minecraft:damage_taken":60001300,"minecraft:walk_one_cm":22359,"minecraft:leave_game":2,"minecraft:sprint_one_cm":20900,"minecraft:play_time":9539,"minecraft:time_since_death":1628,"minecraft:walk_on_water_one_cm":3356,"minecraft:deaths":14,"minecraft:sneak_time":13,"minecraft:total_world_time":9539,"minecraft:walk_under_water_one_cm":3865,"minecraft:swim_one_cm":10595,"minecraft:fall_one_cm":17433,"minecraft:fly_one_cm":154932}},"DataVersion":2975}
